Refactor folder structure of roles & groups pages
