+++ apps/developer-portal/src/pages/groups/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,20 +16,4 @@
  * under the License.
  */
 
-export * from "./applications";
-export * from "./certificates";
-export * from "./claims";
-export * from "./configurations";
-export * from "./email-templates";
-export * from "./customize";
-export * from "./errors";
-export * from "./user-stores";
 export * from "./groups";
-export * from "./home";
-export * from "./privacy";
-export * from "./users";
-export * from "./roles";
-export * from "./identity-providers";
-export * from "./identity-provider-edit";
-export * from "./identity-provider-template";
-export * from "./user-stores";
